+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/provider/foundationdb/FDBMetaDataStoreTest.java
@@ -53,6 +53,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 
+import javax.annotation.Nonnull;
 import java.util.Arrays;
 import java.util.List;
 
@@ -63,6 +64,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNotSame;
+import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -694,4 +696,584 @@ public class FDBMetaDataStoreTest extends FDBTestBase {
             }
         }
     }
+
+    private void addRecordType(@Nonnull DescriptorProtos.DescriptorProto newRecordType, @Nonnull KeyExpression primaryKey) {
+        metaDataStore.mutateMetaData(metaDataProto -> MetaDataProtoEditor.addRecordType(metaDataProto, newRecordType),
+                recordMetaDataBuilder -> {
+                    recordMetaDataBuilder.getRecordType(newRecordType.getName()).setPrimaryKey(primaryKey);
+                    recordMetaDataBuilder.getRecordType(newRecordType.getName()).setSinceVersion(recordMetaDataBuilder.getVersion());
+                });
+    }
+
+    private void deprecateRecordType(@Nonnull String recordType) {
+        metaDataStore.mutateMetaData((metaDataProto) -> MetaDataProtoEditor.deprecateRecordType(metaDataProto, recordType));
+    }
+
+    private void addField(@Nonnull String recordType, @Nonnull DescriptorProtos.FieldDescriptorProto field) {
+        metaDataStore.mutateMetaData((metaDataProto) -> MetaDataProtoEditor.addField(metaDataProto, recordType, field));
+    }
+
+    private void deprecateField(@Nonnull String recordType, @Nonnull String fieldName) {
+        metaDataStore.mutateMetaData((metaDataProto) -> MetaDataProtoEditor.deprecateField(metaDataProto, recordType, fieldName));
+    }
+
+    @Test
+    public void recordTypes() {
+        int version;
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            RecordMetaData metaData = RecordMetaData.build(TestRecords1Proto.getDescriptor());
+            version = metaData.getVersion();
+            metaDataStore.saveRecordMetaData(metaData);
+            context.commit();
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+        }
+
+        // Add an existing record type.
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertEquals(version , metaDataStore.getRecordMetaData().getVersion());
+            DescriptorProtos.DescriptorProto newRecordType = DescriptorProtos.DescriptorProto.newBuilder()
+                    .setName("MySimpleRecord")
+                    .addField(DescriptorProtos.FieldDescriptorProto.newBuilder()
+                            .setLabel(DescriptorProtos.FieldDescriptorProto.Label.LABEL_OPTIONAL)
+                            .setType(DescriptorProtos.FieldDescriptorProto.Type.TYPE_INT32)
+                            .setName("rec_no")
+                            .setNumber(1))
+                    .build();
+            MetaDataException e = assertThrows(MetaDataException.class, () -> addRecordType(newRecordType, Key.Expressions.field("rec_no")));
+            assertEquals(e.getMessage(), "Record type MySimpleRecord already exists");
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertEquals(version, metaDataStore.getRecordMetaData().getVersion()); // version should not change
+            context.commit();
+        }
+
+        // Add a record type.
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertEquals(version , metaDataStore.getRecordMetaData().getVersion());
+            DescriptorProtos.DescriptorProto newRecordType = DescriptorProtos.DescriptorProto.newBuilder()
+                    .setName("MyNewRecord")
+                    .addField(DescriptorProtos.FieldDescriptorProto.newBuilder()
+                            .setLabel(DescriptorProtos.FieldDescriptorProto.Label.LABEL_OPTIONAL)
+                            .setType(DescriptorProtos.FieldDescriptorProto.Type.TYPE_INT32)
+                            .setName("rec_no")
+                            .setNumber(1))
+                    .build();
+            addRecordType(newRecordType, Key.Expressions.field("rec_no"));
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MyNewRecord"));
+            assertEquals(version + 1 , metaDataStore.getRecordMetaData().getVersion());
+            assertEquals(version + 1 , metaDataStore.getRecordMetaData().getRecordType("MyNewRecord").getSinceVersion().intValue());
+            context.commit();
+        }
+
+        // Deprecate the just-added record type.
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertEquals(version + 1, metaDataStore.getRecordMetaData().getVersion());
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MyNewRecord"));
+            deprecateRecordType("MyNewRecord");
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MyNewRecord"));
+            assertTrue(metaDataStore.getRecordMetaData().getRecordsDescriptor().findMessageTypeByName(RecordMetaDataBuilder.DEFAULT_UNION_NAME).findFieldByName("_MyNewRecord").getOptions().getDeprecated());
+            assertEquals(version + 2 , metaDataStore.getRecordMetaData().getVersion());
+            context.commit();
+        }
+
+        // Deprecate a record type from the original proto.
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertEquals(version + 2, metaDataStore.getRecordMetaData().getVersion());
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            deprecateRecordType(".com.apple.foundationdb.record.test1.MySimpleRecord"); // Record type needs to be fully qualified.
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertTrue(metaDataStore.getRecordMetaData().getRecordsDescriptor().findMessageTypeByName(RecordMetaDataBuilder.DEFAULT_UNION_NAME).findFieldByName("_MySimpleRecord").getOptions().getDeprecated());
+            assertEquals(version + 3 , metaDataStore.getRecordMetaData().getVersion());
+            context.commit();
+        }
+
+        // Deprecate a non-existent record type.
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertEquals(version + 3, metaDataStore.getRecordMetaData().getVersion());
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            MetaDataException e = assertThrows(MetaDataException.class, () -> deprecateRecordType("MyNonExistentRecord"));
+            assertEquals(e.getMessage(), "Record type MyNonExistentRecord not found");
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertEquals(version + 3 , metaDataStore.getRecordMetaData().getVersion());
+            context.commit();
+        }
+    }
+
+    @Test
+    public void unionRecordTypes() {
+        int version;
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            RecordMetaData metaData = RecordMetaData.build(TestRecords1Proto.getDescriptor());
+            version = metaData.getVersion();
+            metaDataStore.saveRecordMetaData(metaData);
+            context.commit();
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+        }
+
+        // Add a record type with default union name.
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertEquals(version, metaDataStore.getRecordMetaData().getVersion());
+            DescriptorProtos.DescriptorProto unionRecordType = DescriptorProtos.DescriptorProto.newBuilder()
+                    .setName(RecordMetaDataBuilder.DEFAULT_UNION_NAME)
+                    .addField(DescriptorProtos.FieldDescriptorProto.newBuilder()
+                            .setLabel(DescriptorProtos.FieldDescriptorProto.Label.LABEL_OPTIONAL)
+                            .setType(DescriptorProtos.FieldDescriptorProto.Type.TYPE_INT32)
+                            .setName("rec_no")
+                            .setNumber(1))
+                    .build();
+            MetaDataException e = assertThrows(MetaDataException.class, () -> addRecordType(unionRecordType, Key.Expressions.field("rec_no")));
+            assertEquals(e.getMessage(), "Adding UNION record type not allowed");
+            assertEquals(version, metaDataStore.getRecordMetaData().getVersion()); // version must remain unchanged
+            context.commit();
+        }
+
+        // Add a record type with non-default union name but union usage.
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertEquals(version, metaDataStore.getRecordMetaData().getVersion());
+            DescriptorProtos.MessageOptions.Builder unionMessageOptions = DescriptorProtos.MessageOptions.newBuilder()
+                    .setExtension(RecordMetaDataOptionsProto.record,
+                            RecordMetaDataOptionsProto.RecordTypeOptions.newBuilder()
+                                    .setUsage(RecordMetaDataOptionsProto.RecordTypeOptions.Usage.UNION).build());
+            DescriptorProtos.DescriptorProto unionRecordType = DescriptorProtos.DescriptorProto.newBuilder()
+                    .setName("NonDefaultUnionRecord")
+                    .addField(DescriptorProtos.FieldDescriptorProto.newBuilder()
+                            .setLabel(DescriptorProtos.FieldDescriptorProto.Label.LABEL_OPTIONAL)
+                            .setType(DescriptorProtos.FieldDescriptorProto.Type.TYPE_INT32)
+                            .setName("rec_no")
+                            .setNumber(1))
+                    .setOptions(unionMessageOptions)
+                    .build();
+            MetaDataException e = assertThrows(MetaDataException.class, () -> addRecordType(unionRecordType, Key.Expressions.field("rec_no")));
+            assertEquals(e.getMessage(), "Adding UNION record type not allowed");
+            assertEquals(version, metaDataStore.getRecordMetaData().getVersion()); // version must remain unchanged
+            context.commit();
+        }
+
+        // Deprecate the record type union.
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertEquals(version, metaDataStore.getRecordMetaData().getVersion());
+            MetaDataException e = assertThrows(MetaDataException.class, () -> deprecateRecordType(RecordMetaDataBuilder.DEFAULT_UNION_NAME));
+            assertEquals(e.getMessage(), "Cannot deprecate the union");
+            assertEquals(version, metaDataStore.getRecordMetaData().getVersion()); // version must remain unchanged
+            context.commit();
+        }
+    }
+
+    @Test
+    public void nonDefaultUnionRecordTypes() {
+        int version;
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            RecordMetaDataBuilder metaDataBuilder = RecordMetaData.newBuilder().setRecords(TestRecords3Proto.getDescriptor());
+            metaDataBuilder.getOnlyRecordType().setPrimaryKey(Key.Expressions.concatenateFields("parent_path", "child_name"));
+            RecordMetaData metaData = metaDataBuilder.getRecordMetaData();
+            version = metaData.getVersion();
+            metaDataStore.saveRecordMetaData(metaData);
+            context.commit();
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MyHierarchicalRecord"));
+        }
+
+        // Add a record type with default union name to a record meta-data that has a non-default union name.
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MyHierarchicalRecord"));
+            assertEquals(version, metaDataStore.getRecordMetaData().getVersion());
+            DescriptorProtos.DescriptorProto unionRecordType = DescriptorProtos.DescriptorProto.newBuilder()
+                    .setName(RecordMetaDataBuilder.DEFAULT_UNION_NAME)
+                    .addField(DescriptorProtos.FieldDescriptorProto.newBuilder()
+                            .setLabel(DescriptorProtos.FieldDescriptorProto.Label.LABEL_OPTIONAL)
+                            .setType(DescriptorProtos.FieldDescriptorProto.Type.TYPE_INT32)
+                            .setName("rec_no")
+                            .setNumber(1))
+                    .build();
+            MetaDataException e = assertThrows(MetaDataException.class, () -> addRecordType(unionRecordType, Key.Expressions.field("rec_no")));
+            assertEquals(e.getMessage(), "Adding UNION record type not allowed");
+            assertEquals(version, metaDataStore.getRecordMetaData().getVersion()); // version must remain unchanged
+            context.commit();
+        }
+
+        // Add a second record type with non-default union name but union usage.
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MyHierarchicalRecord"));
+            assertEquals(version, metaDataStore.getRecordMetaData().getVersion());
+            DescriptorProtos.MessageOptions.Builder unionMessageOptions = DescriptorProtos.MessageOptions.newBuilder()
+                    .setExtension(RecordMetaDataOptionsProto.record,
+                            RecordMetaDataOptionsProto.RecordTypeOptions.newBuilder()
+                                    .setUsage(RecordMetaDataOptionsProto.RecordTypeOptions.Usage.UNION).build());
+            DescriptorProtos.DescriptorProto unionRecordType = DescriptorProtos.DescriptorProto.newBuilder()
+                    .setName("SecondNonDefaultUnionRecord")
+                    .addField(DescriptorProtos.FieldDescriptorProto.newBuilder()
+                            .setLabel(DescriptorProtos.FieldDescriptorProto.Label.LABEL_OPTIONAL)
+                            .setType(DescriptorProtos.FieldDescriptorProto.Type.TYPE_INT32)
+                            .setName("rec_no")
+                            .setNumber(1))
+                    .setOptions(unionMessageOptions)
+                    .build();
+            MetaDataException e = assertThrows(MetaDataException.class, () -> addRecordType(unionRecordType, Key.Expressions.field("rec_no")));
+            assertEquals(e.getMessage(), "Adding UNION record type not allowed");
+            assertEquals(version, metaDataStore.getRecordMetaData().getVersion()); // version must remain unchanged
+            context.commit();
+        }
+
+        // Deprecate the record type union with non-default name.
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MyHierarchicalRecord"));
+            assertEquals(version, metaDataStore.getRecordMetaData().getVersion());
+            MetaDataException e = assertThrows(MetaDataException.class, () -> deprecateRecordType("UnionDescriptor"));
+            assertEquals(e.getMessage(), "Cannot deprecate the union");
+            assertEquals(version, metaDataStore.getRecordMetaData().getVersion()); // version must remain unchanged
+            context.commit();
+        }
+
+        // Add a record type to a meta-data that has non-default union name.
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertEquals(version , metaDataStore.getRecordMetaData().getVersion());
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MyHierarchicalRecord"));
+            DescriptorProtos.DescriptorProto newRecordType = DescriptorProtos.DescriptorProto.newBuilder()
+                    .setName("MyNewRecord")
+                    .addField(DescriptorProtos.FieldDescriptorProto.newBuilder()
+                            .setLabel(DescriptorProtos.FieldDescriptorProto.Label.LABEL_OPTIONAL)
+                            .setType(DescriptorProtos.FieldDescriptorProto.Type.TYPE_INT32)
+                            .setName("rec_no")
+                            .setNumber(1))
+                    .build();
+            addRecordType(newRecordType, Key.Expressions.field("rec_no"));
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MyHierarchicalRecord"));
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MyNewRecord"));
+            assertEquals(version + 1 , metaDataStore.getRecordMetaData().getVersion());
+            assertEquals(version + 1 , metaDataStore.getRecordMetaData().getRecordType("MyNewRecord").getSinceVersion().intValue());
+            context.commit();
+        }
+
+        // Deprecate a record type from a meta-data that has non-default union name.
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertEquals(version + 1, metaDataStore.getRecordMetaData().getVersion());
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MyHierarchicalRecord"));
+            deprecateRecordType(".com.apple.foundationdb.record.test3.MyHierarchicalRecord"); // Record type needs to be fully qualified.
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MyHierarchicalRecord"));
+            assertTrue(metaDataStore.getRecordMetaData().getRecordsDescriptor().findMessageTypeByName("UnionDescriptor").findFieldByName("_MyHierarchicalRecord").getOptions().getDeprecated());
+            assertEquals(version + 2, metaDataStore.getRecordMetaData().getVersion());
+            context.commit();
+        }
+    }
+
+    private void addNestedRecordType(DescriptorProtos.DescriptorProto newRecordType) {
+        metaDataStore.mutateMetaData((metaDataProto) -> MetaDataProtoEditor.addNestedRecordType(metaDataProto, newRecordType));
+    }
+
+    @Test
+    public void nestedRecordTypes() {
+        int version;
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            RecordMetaData metaData = RecordMetaData.build(TestRecords1Proto.getDescriptor());
+            version = metaData.getVersion();
+            metaDataStore.saveRecordMetaData(metaData);
+            context.commit();
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+        }
+
+        // Adding an existing record type should fail
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertEquals(version , metaDataStore.getRecordMetaData().getVersion());
+            DescriptorProtos.MessageOptions.Builder nestedMessageOptions = DescriptorProtos.MessageOptions.newBuilder()
+                    .setExtension(RecordMetaDataOptionsProto.record,
+                            RecordMetaDataOptionsProto.RecordTypeOptions.newBuilder()
+                                    .setUsage(RecordMetaDataOptionsProto.RecordTypeOptions.Usage.NESTED).build());
+            DescriptorProtos.DescriptorProto newRecordType = DescriptorProtos.DescriptorProto.newBuilder()
+                    .setName("MySimpleRecord")
+                    .addField(DescriptorProtos.FieldDescriptorProto.newBuilder()
+                            .setLabel(DescriptorProtos.FieldDescriptorProto.Label.LABEL_OPTIONAL)
+                            .setType(DescriptorProtos.FieldDescriptorProto.Type.TYPE_INT32)
+                            .setName("rec_no")
+                            .setNumber(1))
+                    .setOptions(nestedMessageOptions)
+                    .build();
+            MetaDataException e = assertThrows(MetaDataException.class, () -> addNestedRecordType(newRecordType));
+            assertEquals(e.getMessage(), "Record type MySimpleRecord already exists");
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertEquals(version, metaDataStore.getRecordMetaData().getVersion()); // version should not change
+            context.commit();
+        }
+
+        // Add a nested record type.
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertEquals(version, metaDataStore.getRecordMetaData().getVersion());
+            DescriptorProtos.MessageOptions.Builder nestedMessageOptions = DescriptorProtos.MessageOptions.newBuilder()
+                    .setExtension(RecordMetaDataOptionsProto.record,
+                            RecordMetaDataOptionsProto.RecordTypeOptions.newBuilder()
+                                    .setUsage(RecordMetaDataOptionsProto.RecordTypeOptions.Usage.NESTED).build());
+            DescriptorProtos.DescriptorProto newRecordType = DescriptorProtos.DescriptorProto.newBuilder()
+                    .setName("MyNewNestedRecord")
+                    .addField(DescriptorProtos.FieldDescriptorProto.newBuilder()
+                            .setLabel(DescriptorProtos.FieldDescriptorProto.Label.LABEL_OPTIONAL)
+                            .setType(DescriptorProtos.FieldDescriptorProto.Type.TYPE_INT32)
+                            .setName("rec_no")
+                            .setNumber(1))
+                    .setOptions(nestedMessageOptions)
+                    .build();
+
+            // Use addRecordType should fail
+            MetaDataException e = assertThrows(MetaDataException.class, () -> addRecordType(newRecordType, Key.Expressions.field("rec_no")));
+            assertEquals(e.getMessage(), "Use addNestedRecordType for adding NESTED record types");
+
+            // Use addNestedRecordType should succeed
+            addNestedRecordType(newRecordType);
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            e = assertThrows(MetaDataException.class, () -> metaDataStore.getRecordMetaData().getRecordType("MyNewNestedRecord"));
+            assertEquals(e.getMessage(), "Unknown record type MyNewNestedRecord");
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordsDescriptor().findMessageTypeByName("MyNewNestedRecord"));
+            assertNull(metaDataStore.getRecordMetaData().getRecordsDescriptor()
+                    .findMessageTypeByName(RecordMetaDataBuilder.DEFAULT_UNION_NAME)
+                    .findFieldByName("_MyNewNestedRecord"));
+            assertEquals(version + 1, metaDataStore.getRecordMetaData().getVersion());
+
+            // addNestedRecordType is not idempotent!
+            e = assertThrows(MetaDataException.class, () -> addNestedRecordType(newRecordType));
+            assertEquals(e.getMessage(), "Record type MyNewNestedRecord already exists");
+            assertEquals(version + 1, metaDataStore.getRecordMetaData().getVersion());
+            context.commit();
+        }
+
+        // Add nested type as a field
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordsDescriptor().findMessageTypeByName("MyNewNestedRecord"));
+            assertEquals(version + 1 , metaDataStore.getRecordMetaData().getVersion());
+
+            DescriptorProtos.FieldDescriptorProto field = DescriptorProtos.FieldDescriptorProto.newBuilder()
+                    .setName("newField")
+                    .setType(DescriptorProtos.FieldDescriptorProto.Type.TYPE_MESSAGE)
+                    .setTypeName("MyNewNestedRecord")
+                    .setLabel(DescriptorProtos.FieldDescriptorProto.Label.LABEL_OPTIONAL)
+                    .setNumber(10)
+                    .build();
+            addField("MySimpleRecord", field);
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordsDescriptor().findMessageTypeByName("MySimpleRecord").findFieldByName("newField"));
+            assertEquals(metaDataStore.getRecordMetaData().getRecordsDescriptor().findMessageTypeByName("MyNewNestedRecord"),
+                    metaDataStore.getRecordMetaData().getRecordsDescriptor().findMessageTypeByName("MySimpleRecord").findFieldByName("newField").getMessageType());
+            assertEquals(version + 2 , metaDataStore.getRecordMetaData().getVersion());
+            context.commit();
+        }
+
+        // Deprecate a nested record type.
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertEquals(version + 2 , metaDataStore.getRecordMetaData().getVersion());
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordsDescriptor().findMessageTypeByName("MyNewNestedRecord"));
+            deprecateField("MySimpleRecord", "newField");
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertTrue(metaDataStore.getRecordMetaData().getRecordsDescriptor()
+                    .findMessageTypeByName("MySimpleRecord")
+                    .findFieldByName("newField")
+                    .getOptions()
+                    .getDeprecated());
+            assertEquals(version + 3, metaDataStore.getRecordMetaData().getVersion());
+            context.commit();
+        }
+    }
+
+    @Test
+    public void fields() {
+        int version;
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            RecordMetaData metaData = RecordMetaData.build(TestRecords1Proto.getDescriptor());
+            version = metaData.getVersion();
+            metaDataStore.saveRecordMetaData(metaData);
+            context.commit();
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+        }
+
+        // Add a new field
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertEquals(version , metaDataStore.getRecordMetaData().getVersion());
+            DescriptorProtos.FieldDescriptorProto field = DescriptorProtos.FieldDescriptorProto.newBuilder()
+                    .setName("newField")
+                    .setType(DescriptorProtos.FieldDescriptorProto.Type.TYPE_INT32)
+                    .setLabel(DescriptorProtos.FieldDescriptorProto.Label.LABEL_OPTIONAL)
+                    .setNumber(10)
+                    .build();
+            addField("MySimpleRecord", field);
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordsDescriptor().findMessageTypeByName("MySimpleRecord").findFieldByName("newField"));
+            assertEquals(version + 1 , metaDataStore.getRecordMetaData().getVersion());
+
+            // Add it again should fail
+            MetaDataException e = assertThrows(MetaDataException.class, () -> addField("MySimpleRecord", field));
+            assertEquals(e.getMessage(), "Field newField already exists in record type MySimpleRecord");
+            context.commit();
+        }
+
+        // Add a field with non-existent record type
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertEquals(version + 1, metaDataStore.getRecordMetaData().getVersion());
+            DescriptorProtos.FieldDescriptorProto field = DescriptorProtos.FieldDescriptorProto.newBuilder()
+                    .setName("newFieldWithNonExistentRecordType")
+                    .setType(DescriptorProtos.FieldDescriptorProto.Type.TYPE_MESSAGE)
+                    .setTypeName("NonExistentType")
+                    .setLabel(DescriptorProtos.FieldDescriptorProto.Label.LABEL_OPTIONAL)
+                    .setNumber(10)
+                    .build();
+            MetaDataException e = assertThrows(MetaDataException.class, () -> addField("MySimpleRecord", field));
+            assertEquals(e.getMessage(), "Error converting from protobuf");
+            assertEquals(version + 1 , metaDataStore.getRecordMetaData().getVersion());
+            context.commit();
+        }
+
+        // Deprecate field should fail if record type or field does not exist
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertEquals(version + 1, metaDataStore.getRecordMetaData().getVersion());
+
+            MetaDataException e = assertThrows(MetaDataException.class, () -> deprecateField("NonExistentRecordType", "field"));
+            assertEquals(e.getMessage(), "Record type NonExistentRecordType does not exist");
+
+            e = assertThrows(MetaDataException.class, () -> deprecateField("MySimpleRecord", "nonExistentField"));
+            assertEquals(e.getMessage(), "Field nonExistentField not found in record type MySimpleRecord");
+            context.commit();
+        }
+
+        // Deprecate a field
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordsDescriptor().findMessageTypeByName("MySimpleRecord").findFieldByName("num_value_2"));
+            assertFalse(metaDataStore.getRecordMetaData()
+                    .getRecordsDescriptor()
+                    .findMessageTypeByName("MySimpleRecord")
+                    .findFieldByName("num_value_2")
+                    .getOptions()
+                    .hasDeprecated());
+            assertEquals(version + 1 , metaDataStore.getRecordMetaData().getVersion());
+            deprecateField("MySimpleRecord", "num_value_2");
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertTrue(metaDataStore.getRecordMetaData()
+                    .getRecordsDescriptor()
+                    .findMessageTypeByName("MySimpleRecord")
+                    .findFieldByName("num_value_2")
+                    .getOptions()
+                    .getDeprecated());
+            assertEquals(version + 2 , metaDataStore.getRecordMetaData().getVersion());
+            context.commit();
+        }
+
+        // Deprecate the newly added field
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordsDescriptor().findMessageTypeByName("MySimpleRecord").findFieldByName("newField"));
+            assertFalse(metaDataStore.getRecordMetaData()
+                    .getRecordsDescriptor()
+                    .findMessageTypeByName("MySimpleRecord")
+                    .findFieldByName("newField")
+                    .getOptions()
+                    .hasDeprecated());
+            assertEquals(version + 2 , metaDataStore.getRecordMetaData().getVersion());
+            deprecateField("MySimpleRecord", "newField");
+            assertNotNull(metaDataStore.getRecordMetaData().getRecordType("MySimpleRecord"));
+            assertTrue(metaDataStore.getRecordMetaData()
+                    .getRecordsDescriptor()
+                    .findMessageTypeByName("MySimpleRecord")
+                    .findFieldByName("newField")
+                    .getOptions()
+                    .getDeprecated());
+            assertEquals(version + 3 , metaDataStore.getRecordMetaData().getVersion());
+            context.commit();
+        }
+    }
+
+    @Test
+    public void updateSchemaOptions() {
+        int version;
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            RecordMetaData metaData = RecordMetaData.build(TestRecords1Proto.getDescriptor());
+            metaDataStore.saveRecordMetaData(metaData);
+            version = metaDataStore.getRecordMetaData().getVersion();
+            context.commit();
+        }
+
+        // Unset store record versions.
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertTrue(metaDataStore.getRecordMetaData().isStoreRecordVersions());
+            metaDataStore.updateStoreRecordVersions(false);
+            context.commit();
+            assertEquals(version + 1 , metaDataStore.getRecordMetaData().getVersion());
+            assertFalse(metaDataStore.getRecordMetaData().isStoreRecordVersions());
+        }
+
+        // Set store record versions.
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertFalse(metaDataStore.getRecordMetaData().isStoreRecordVersions());
+            metaDataStore.updateStoreRecordVersions(true);
+            context.commit();
+            assertEquals(version + 2 , metaDataStore.getRecordMetaData().getVersion());
+            assertTrue(metaDataStore.getRecordMetaData().isStoreRecordVersions());
+        }
+
+        // Enable split long records with default validator. It should fail.
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertFalse(metaDataStore.getRecordMetaData().isSplitLongRecords());
+            MetaDataException e = assertThrows(MetaDataException.class, () -> metaDataStore.enableSplitLongRecords());
+            assertEquals(e.getMessage(), "new meta-data splits long records");
+            context.commit();
+            assertEquals(version + 2 , metaDataStore.getRecordMetaData().getVersion());
+            assertTrue(metaDataStore.getRecordMetaData().isStoreRecordVersions());
+        }
+
+        // Enable split long records
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            metaDataStore.setEvolutionValidator(MetaDataEvolutionValidator.newBuilder().setAllowUnsplitToSplit(true).build());
+            assertFalse(metaDataStore.getRecordMetaData().isSplitLongRecords());
+            metaDataStore.enableSplitLongRecords();
+            context.commit();
+            assertEquals(version + 3 , metaDataStore.getRecordMetaData().getVersion());
+            assertTrue(metaDataStore.getRecordMetaData().isStoreRecordVersions());
+        }
+
+        try (FDBRecordContext context = fdb.openContext()) {
+            openMetaDataStore(context);
+            assertTrue(metaDataStore.getRecordMetaData().isStoreRecordVersions());
+            assertEquals(version + 3 , metaDataStore.getRecordMetaData().getVersion());
+            context.commit();
+        }
+    }
+
 }